pyproject.toml
# pyproject.toml

[tool.poetry]
name = "rag-agent-framework"
version = "0.1.0"
description = "A modular, production-grade RAG + Agent framework."
authors = ["Alex Vo ma@alex-vo.com"]
readme = "README.md"
packages = [{include = "rag_agent_framework", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11, <=3.13"

# Core LangChain & CrewAI - Aligned for LangChain v0.2.x
langchain = "^0.2.5"
langchain-openai = "^0.1.8"
langchain-community = "^0.2.5"
crewai = "^0.130.0"

# Vector DB (Chroma included but code will use Qdrant)
chromadb = "^0.5.3"
qdrant-client = "^1.7.0"

# LLM Support
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
ollama = "^0.2.1"

# Data Ingestion
pypdf = "^4.2.0"
beautifulsoup4 = "^4.12.3"

# Backend & Frontend
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
streamlit = "^1.35.0"
requests = "^2.32.0"

# Tooling
duckduckgo-search = "^6.1.8"
langchain-qdrant = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"